pyproject.toml
[project]
name = "pool-exporter"
version = "0.0.1"
description = "Data exporter for Zürich swimming pool occupancy"
authors = [
  { name = "Your Name", email = "you@example.com" },
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "websockets >=15.0.1",
  "prometheus-client >=0.21.1",
  "pyyaml >=6.0.2",
  "aiohttp >=3.11.18",
]

[project.scripts]
pool-exporter = "pool_exporter.exporter:run"

[dependency-groups]
dev = [
  "mypy>=1.17.0",
  "pytest>=8.4.1",
  "pytest-asyncio>=1.3.0",
  "pytest-mock>=3.15.1",
  "requests>=2.32.4",
  "types-pyyaml>=6.0.12.20250402",
  "types-requests>=2.32.4.20250611",
]

[build-system]
# uv.lock explicitly excludes build-system dependencies, so they aren't pinned.
# Dependabot fails with opaque "content not changed" errors if this version is too old.
# Following official docs:
# https://docs.astral.sh/uv/concepts/build-backend/#choosing-a-build-backend
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[tool.mypy]
strict = true

[tool.ruff]
lint.select = [
  "E", # pycodestyle
  "F", # Pyflakes
  "UP", # pyupgrade
  "B", # flake8-bugbear
  "SIM", # flake8-simplify
  "I", # isort
  "C90", # mccabe complexity
]
